@@ -1,15 +1,15 @@
 apply plugin: 'com.android.application'
 android {
     compileOptions.incremental = false
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.koohpar.dstrbt"
         minSdkVersion 17
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 3
         versionName "1.2.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-        buildToolsVersion '28.0.3'
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        buildToolsVersion '29.0.0'
         multiDexEnabled true
     }
     dataBinding {
