@@ -16,7 +16,7 @@
 
 package com.koohpar.dstrbt.ui.base;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 public abstract class BaseViewHolder extends RecyclerView.ViewHolder {
